Eta refactored (#264)
* escape

* Separated eta.go file into eta_math file and manager file

* Enables /eta endpoint

* logs all calculated etas to the local storage for the browswer

* fixed build error with writejson

* Delete settings.json

* Update go.mod
